site: move HERO_H1/SITE_TITLE into HeroSection, drop copy.ts
