+++ site/vite.config.ts
@@ -2,7 +2,7 @@ import { defineConfig, type Plugin } from "vite";
 import preact from "@preact/preset-vite";
 import tailwindcss from "@tailwindcss/vite";
 import { resolve } from "node:path";
-import { SITE_TITLE } from "./src/copy";
+import { SITE_TITLE } from "./src/sections/HeroSection";
 
 // Keep <title> in sync with HeroSection's H1. The constant lives in
 // src/copy.ts so HeroSection imports it directly; this plugin rewrites
